Working support for i686 physmem.
There are a few hacks involved in this.  This includes the addition of the bootlist page list, which contains pages used at boot time (for things like grub structures) than can be freed at the end of boot.

Since mutex is not complete/functional, there is actually no smp support in this code due to no locking at all.
Planned is having per-cpu page lists that can trade pages back and forth when there is memory pressure.